+++ src/constants/poe.ts
@@ -1,0 +1,5 @@
+export enum ProofOfExistenceMode {
+  ENABLED_NOT_THROWABLE,
+  STRICT,
+  DISABLED
+}