+++ src/services/verifiable-credentials/ddcc.format.ts
@@ -1,5 +1,11 @@
 import { Type } from 'class-transformer';
-import { IsDefined, IsNumber, IsString, ValidateNested } from 'class-validator';
+import {
+  IsDefined,
+  IsNumber,
+  IsOptional,
+  IsString,
+  ValidateNested
+} from 'class-validator';
 
 export class Identifier {
   @IsString()
@@ -10,44 +16,67 @@ export class Issuer {
   identifier!: Identifier;
 }
 
+export class Period {
+  @IsString()
+  start!: string;
+  @IsString()
+  end!: string;
+}
+
 export class DDCCCertificate {
-  // period .. omitted
+  @IsOptional()
+  @Type(() => Period)
+  period!: Period;
   @Type(() => Identifier)
   hcid!: Identifier;
-  // version .. omitted
+  @IsString()
+  version!: string;
   @Type(() => Issuer)
   issuer!: Issuer;
 }
 
-export class Vaccine {
-  @IsString()
-  code!: string;
-}
-export class Country {
-  @IsString()
-  code!: string;
-}
-
-export class Brand {
+export class CodeSystem {
   @IsString()
   code!: string;
 }
 
 export class Vaccination {
-  @Type(() => Vaccine)
-  vaccine!: Vaccine;
+  @Type(() => CodeSystem)
+  vaccine!: CodeSystem;
   @IsString()
   date!: string;
   @IsNumber()
   dose!: number;
-  @Type(() => Country)
-  country!: Country;
+  @Type(() => CodeSystem)
+  country!: CodeSystem;
   @IsString()
+  @IsOptional()
   centre!: string;
-  @Type(() => Brand)
-  brand!: Brand;
+  @IsString()
+  @IsOptional()
+  nextDose!: string;
+  @Type(() => CodeSystem)
+  brand!: CodeSystem;
   @IsString()
   lot!: string;
+  @IsOptional()
+  @Type(() => CodeSystem)
+  @ValidateNested()
+  maholder!: CodeSystem;
+  @IsOptional()
+  @Type(() => CodeSystem)
+  @ValidateNested()
+  disease!: CodeSystem;
+  @IsOptional()
+  @IsString()
+  totalDoses!: string;
+  @IsOptional()
+  @IsString()
+  validFrom!: string;
+  @IsOptional()
+  @Type(() => Identifier)
+  @ValidateNested()
+  practitioner!: Identifier;
 }
 
 export class DDCCFormatValidator {
@@ -60,10 +89,13 @@ export class DDCCFormatValidator {
   @IsString()
   name!: string;
   @IsString()
+  @IsOptional()
   birthDate!: string;
   @IsString()
+  @IsOptional()
   identifier!: string;
   @IsString()
+  @IsOptional()
   sex!: string;
 }